Fix for parsing date format used by some Windows ftp servers.
The code was not correctly parsing dates returned from a Windows ftp
server, causing no files to be reported from the server. The date
format has four digit years and the parsing code expected a two digit
year. This change allows there to be 2 or 4 digit years in the dates.
Task-number: QTBUG-35177

Change-Id: Ibbd7333776f7f998f1cc569a8768a1658bb0a376
Reviewed-by: Peter Hartmann <phartmann@blackberry.com>